refactor: multiple stories
