+++ src/components/Pentagon/__stories__/Playground.stories.tsx
@@ -1,0 +1,3 @@
+import { Template } from './Template.stories'
+
+export const Playground = Template.bind({})
