+++ src/components/Pentagon/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Pentagon from '..'
+
+export const Template: ComponentStory<typeof Pentagon> = args => (
+  <Pentagon {...args} />
+)
