+++ src/components/ProgressBar/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import ProgressBar from '..'
+
+export const Template: ComponentStory<typeof ProgressBar> = args => (
+  <ProgressBar value={40} {...args} />
+)
